Added optional args to function
Function now has features:
- Prints directory and file it is working on
- Optional arg to either overwrite file or skip file if it already exists
- Optional arg to specify which tasks to run or skip
- Optional arg to specify which subjects to run or skip
Also did general bug fixing

+++ rename_data_before_training.py
@@ -24,28 +24,39 @@ def extract_timestamp(filename):
         empty list or false value will default to processing all 
         data in the source directory
 '''
-def hs_rename_files(source_dir, destination_dir, subjects):
+def hs_rename_files(source_dir, destination_dir, subjects=[], overwrite_file=True,
+                    tasks_to_run=["FMA17", "FMA18", "FMA19", "FMA20", "FMA21", 
+                                  "FMA22", "FMA23"], tasks_to_skip=[],
+                    subjects_to_skip=[]):
     if len(subjects) == 0:
         subject_dirs = [dir for dir in os.listdir(source_dir) if 
-                        os.path.isdir(os.path.join(source_dir, dir))]
+                        os.path.isdir(os.path.join(source_dir, dir))
+                        and dir[0:2] not in subjects_to_skip]
     else:
         subject_dirs = [dir for dir in os.listdir(source_dir) if 
                         os.path.isdir(os.path.join(source_dir, dir)) and 
                         dir[0:2] in subjects]
+    
+    tasks = list(set(tasks_to_run) - set(tasks_to_skip))
+
+    # Prepare destination directory
+    if not os.path.isdir(os.path.join(destination_dir, "HS_data")):
+        os.makedirs(os.path.join(destination_dir, "HS_data"))
 
     # Extract subject_id from dir name and create destination path
     for subject_dir in subject_dirs:
         subject_dir_path = os.path.join(source_dir, subject_dir)
         subject_id = subject_dir[0:2]
-        new_subject_dir = os.path.join(destination_dir, subject_id)
+        new_subject_dir = os.path.join(destination_dir, "HS_data", subject_id)
 
         # Create destination folder if it does not exist
         if not os.path.isdir(new_subject_dir):
             os.makedirs(new_subject_dir)
-            for num in range(17, 24):  # Create FMA17 to FMA23 folders
-                os.makedirs(os.path.join(new_subject_dir, f"FMA{num}"))
+        for task in tasks:  # Create FMA folders
+            if not os.path.isdir(os.path.join(new_subject_dir, task)):
+                os.makedirs(os.path.join(new_subject_dir, task))
 
-
+        print(f"SUBJECT: {subject_dir}")
         score_dirs = os.listdir(subject_dir_path)
         for score_dir in score_dirs:
             
@@ -63,6 +74,9 @@ def hs_rename_files(source_dir, destination_dir, subjects):
             for file in csv_files:
                 new_filename_parts = file.split("_")
                 fma_num = new_filename_parts[0][0:5]
+                
+                if fma_num not in tasks:
+                    continue
 
                 if not fma_num[3:5].isdigit():
                     continue  # Skip files that don't have the FMA num
@@ -82,13 +96,16 @@ def hs_rename_files(source_dir, destination_dir, subjects):
                 fma_files[fma_num].append((timestamp, file, score_dir_path, new_filename))
 
             for fma_num, files in fma_files.items():
+                
                 sorted_files = sorted(files, key=lambda x: x[0])
 
                 for trial_num, (_, file, from_dir, filename) in enumerate(sorted_files, start=1):
                     new_filename_parts = filename.split("_")
                     new_filename_parts.insert(5, f"Trial{trial_num}")
                     new_filename = "_".join(new_filename_parts)
-                    shutil.copy(os.path.join(from_dir, file), 
+                    print(f"File: {file}")
+                    if not os.path.exists(os.path.join(new_subject_dir, fma_num, new_filename)) or overwrite_file:
+                        shutil.copy(os.path.join(from_dir, file), 
                                 os.path.join(new_subject_dir, fma_num, new_filename))
 
 ''' 
@@ -103,14 +120,24 @@ def hs_rename_files(source_dir, destination_dir, subjects):
         empty list or false value will default to processing all 
         data in the source directory
 '''
-def sp_rename_files(source_dir, destination_dir, subjects):
-    if not subjects or len(subjects) == 0:
+def sp_rename_files(source_dir, destination_dir, subjects=[], overwrite_file=True,
+                    tasks_to_run=["FMA17", "FMA18", "FMA19", "FMA20", "FMA21", 
+                                  "FMA22", "FMA23"], tasks_to_skip=[],
+                    subjects_to_skip=[]):
+    if len(subjects) == 0:
         subject_dirs = [dir for dir in os.listdir(source_dir) if 
-                        os.path.isdir(os.path.join(source_dir, dir))]
+                        os.path.isdir(os.path.join(source_dir, dir))
+                        and dir not in subjects_to_skip]
     else:
         subject_dirs = [dir for dir in os.listdir(source_dir) if 
                         os.path.isdir(os.path.join(source_dir, dir)) and 
                         dir in subjects]
+    
+    tasks = list(set(tasks_to_run) - set(tasks_to_skip))
+
+    # Prepare destination directory
+    if not os.path.isdir(os.path.join(destination_dir, "SP_data")):
+        os.makedirs(os.path.join(destination_dir, "SP_data"))
     
     # Extract subject_id from dir name and create destination path
     for subject_dir in subject_dirs:
@@ -141,14 +168,17 @@ def sp_rename_files(source_dir, destination_dir, subjects):
                 lefthand[df.iloc[row,0][0:5]] = df.iloc[row, 3]
         fma_scores = {"lefthand": lefthand, "righthand": righthand}
 
-        new_subject_dir = os.path.join(destination_dir, subject_dir)
 
         # Create destination folder if it does not exist
+        new_subject_dir = os.path.join(destination_dir, "SP_data", subject_dir)
+
         if not os.path.isdir(new_subject_dir):
             os.makedirs(new_subject_dir)
-            for num in range(17, 24):  # Create FMA17 to FMA23 folders
-                os.makedirs(os.path.join(new_subject_dir, f"FMA{num}"))
+        for task in tasks:  # Create FMA folders
+            if not os.path.isdir(os.path.join(new_subject_dir, task)):
+                os.makedirs(os.path.join(new_subject_dir, task))
 
+        print(f"SUBJECT: {subject_dir}")
         hand_dirs = os.listdir(subject_dir_path)
         for hand_dir in hand_dirs:
             
@@ -159,12 +189,8 @@ def sp_rename_files(source_dir, destination_dir, subjects):
                 continue
             
             hand_dir_parts = hand_dir.split("_")
-            affection = hand_dir_parts[3] # Extract affection from dir name
-            side = ""
-            if hand_dir_parts[4] == "LeftHand":
-                side = "lefthand"
-            else:
-                side = "righthand"
+            affection = "Affected" if "Affected" in hand_dir_parts else "Unaffected" # Extract affection from dir name
+            side = "lefthand" if "LeftHand" in hand_dir_parts else "righthand"
             
             csv_files = [file for file in sorted(os.listdir(hand_dir_path)) 
                          if file.endswith('.csv')]
@@ -172,6 +198,9 @@ def sp_rename_files(source_dir, destination_dir, subjects):
             for file in csv_files:
                 new_filename_parts = file.split("_")
                 fma_num = new_filename_parts[0][0:5]
+
+                if fma_num not in tasks:
+                    continue
 
                 if not fma_num[3:5].isdigit():
                     continue  # Skip files that don't have the FMA num
@@ -198,27 +227,39 @@ def sp_rename_files(source_dir, destination_dir, subjects):
                     new_filename_parts = filename.split("_")
                     new_filename_parts.insert(-6, f"Trial{trial_num}")
                     new_filename = "_".join(new_filename_parts)
-                    shutil.copy(os.path.join(from_dir, file), 
+                    print(f"File: {file}")
+                    if not os.path.exists(os.path.join(new_subject_dir, fma_num, new_filename)) or overwrite_file:
+                        shutil.copy(os.path.join(from_dir, file), 
                                 os.path.join(new_subject_dir, fma_num, new_filename))
         
 if __name__ == '__main__':
 
-    DATA_TYPE = ["HS"] # "HS" and/or "SP" ; "HS" means healthy subjects - 
-                            # tendendy data AND "SP" means Stroke Patients - actual experiment data 
+    # ==Configurations==
+    DATA_TYPE = ["SP"] # "HS" and/or "SP" ; "HS" means healthy subjects - tendency data
+                            # AND "SP" means Stroke Patients - actual experiment data 
                # ["HS"]
+    overwrite_file = False  # Setting to overwrite files with new file if they already exist
+                            # By default, function automatically overwrites files
+
+    # Specifies which tasks (FMA17-23) to run. If parameter isn't passed, all tasks are run 
+    tasks_to_run = ["FMA17", "FMA18", "FMA19", "FMA20" "FMA21", "FMA22", "FMA23"]
+
+    # Specifies any tasks (FMA17-23) to skip. If parameter isn't passed, none are skipped
+    tasks_to_skip = []
+
     # ===HS=== 
     # original input source directory   
     source_hs = r'c:\Users\FireF\Box\FMA_Experiment_Share\FMAhand-sensor\HS-tendency-data'
-    subjects_hs = ["VV"]    # subjects that want to run, passing nothing defaults to all
+    subjects_hs = ["VV"]    # Specifies subjects to run. If parameter isn't passed, all subjects are run
     
     # ===SP===
     source_sp = r'C:\Users\FireF\Box\FMA_Experiment_Share\SP'
-    subjects_sp = ["ID12-DBML-E250212"]    # subjects that want to run, passing nothing defaults to all
+    subjects_sp = ["ID12-DBML-E250212"]    # Specifies subjects to run. If parameter isn't passed, all subjects are run
 
     # output directory 
-    destination = r'c:\Users\FireF\Box\FMA_Experiment_Share\FMAhand-sensor\data_before_training'
+    destination = r'c:\Users\FireF\Box\FMA_Experiment_Share\FMAhand-sensor\data_for_processing'
     
     if "HS" in DATA_TYPE:
-        hs_rename_files(source_hs, destination, subjects_hs)
+        hs_rename_files(source_hs, destination)
     if "SP" in DATA_TYPE:
-        sp_rename_files(source_sp, destination, subjects_sp)
+        sp_rename_files(source_sp, destination)
